Update extended to not include uri version
